refactor: move scaling modules under package
