+++ ai_trading/capital_scaling.py
