+++ ai_trading/trade_logic.py
@@ -3,7 +3,7 @@
 import random
 import metrics_logger
 from logger import get_logger
-from capital_scaling import drawdown_adjusted_kelly
+from ai_trading.capital_scaling import drawdown_adjusted_kelly
 
 log = get_logger(__name__)
