+++ tests/test_bot_extended.py
@@ -41,7 +41,7 @@ mods = [
     "pybreaker",
     "ratelimit",
     "trade_execution",
-    "capital_scaling",
+    "ai_trading.capital_scaling",
     "strategy_allocator",
 ]
 for name in mods:
@@ -50,7 +50,7 @@ for name in mods:
 
 # Provide a minimal CapitalScalingEngine so bot imports succeed and
 # tests can call ``update`` without errors.
-if "capital_scaling" in sys.modules:
+if "ai_trading.capital_scaling" in sys.modules:
     class _CapScaler:
         def __init__(self, *a, **k):
             pass
@@ -58,7 +58,7 @@ if "capital_scaling" in sys.modules:
         def update(self, *a, **k):
             pass
 
-    sys.modules["capital_scaling"].CapitalScalingEngine = _CapScaler
+    sys.modules["ai_trading.capital_scaling"].CapitalScalingEngine = _CapScaler
 
 sys.modules.setdefault("yfinance", types.ModuleType("yfinance"))
 if "pandas_market_calendars" in sys.modules:
