+++ tests/test_kelly_drawdown_taper.py
@@ -1,5 +1,5 @@
 import pytest
-from capital_scaling import drawdown_adjusted_kelly
+from ai_trading.capital_scaling import drawdown_adjusted_kelly
 
 
 def test_drawdown_adjusted_kelly_basic():
